requirements.txt
PyQt5==5.15.9
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.10.3
pytesseract==0.3.10
pdfrw==0.4
reportlab==4.0.4
pdf2image==1.17.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
